[Ref] J2B: Silence GCC -Wconversion warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19265 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/load_j2b.cpp
@@ -1041,7 +1041,7 @@ bool CSoundFile::ReadJ2B(FileReader &file, ModLoadingFlags loadFlags)
 		Bytef buffer[mpt::IO::BUFFERSIZE_TINY];
 		uint32 readSize = std::min(static_cast<uint32>(sizeof(buffer)), remainRead);
 		file.ReadRaw(mpt::span(buffer, readSize));
-		crc = crc32(crc, buffer, readSize);
+		crc = static_cast<uint32>(crc32(crc, buffer, readSize));
 
 		strm.avail_in = readSize;
 		strm.next_in = buffer;
